% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_arrow_ds}
\alias{as_arrow_ds}
\title{Coerce an mrgsimsds object to an arrow data set}
\usage{
as_arrow_ds(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object or a list of mrgsimsds objects.}

\item{...}{not used.}
}
\value{
An 'Apache' 'Arrow' \link[arrow:Dataset]{arrow::Dataset} object.
}
\description{
Extracts the underlying \link[arrow:Dataset]{arrow::Dataset} from an mrgsimsds object, allowing
you to work directly with the Arrow API or pass the dataset to other
Arrow-aware tools. For a list, only \code{mrgsimsds} elements are retained and
a single dataset spanning all their files is returned.
}
\examples{
mod <- house_ds(end = 5)

out <- mrgsim_ds(mod, events = ev(amt = 100))

as_arrow_ds(out)

}
