% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_duckdb_ds}
\alias{as_duckdb_ds}
\title{Coerce an mrgsimsds object to a DuckDB table}
\usage{
as_duckdb_ds(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object or a list of mrgsimsds objects.}

\item{...}{passed to \code{\link[=as_arrow_ds]{as_arrow_ds()}}.}
}
\value{
A \code{tbl} of the simulated data in DuckDB; see \code{\link[arrow:to_duckdb]{arrow::to_duckdb()}}.
}
\description{
The conversion is handled by \code{\link[=as_arrow_ds]{as_arrow_ds()}}.
}
\examples{
mod <- house_ds(end = 5)

out <- mrgsim_ds(mod, events = ev(amt = 100))

if(requireNamespace("duckdb")) {
  as_duckdb_ds(out)
}

}
\seealso{
\code{\link[=as_arrow_ds]{as_arrow_ds()}}
}
