% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-ds.R
\name{as_mrgsim_ds}
\alias{as_mrgsim_ds}
\title{Coerce an mrgsims object to 'Apache' 'Arrow'-backed mrgsimsds object}
\usage{
as_mrgsim_ds(x, verbose = FALSE, gc = TRUE)
}
\arguments{
\item{x}{an mrgsims object.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{gc}{initial gc setting; if \code{TRUE}, a finalizer function will attempt
to remove files once the object is out of scope. This value is not locked:
\code{\link[=move_ds]{move_ds()}} and \code{\link[=save_ds]{save_ds()}} will automatically adjust gc based on whether
the files remain in the output directory where they were written. To lock the
gc setting and prevent automatic adjustment, call \code{\link[=gc_ds]{gc_ds()}} after creation.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Converts the output of \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}} to an \code{mrgsimsds} object by
writing the simulation data to a parquet file in the output directory
attached to the model object (\code{tempdir()} by default; see \code{\link[=mread_ds]{mread_ds()}} and
\code{\link[=set_ds_dir]{set_ds_dir()}}). Files in the output directory are auto-deleted on garbage
collection by default. Use \code{\link[=move_ds]{move_ds()}} or \code{\link[=save_ds]{save_ds()}} to relocate files
outside the output directory, which automatically disables gc, or call
\code{\link[=gc_ds]{gc_ds()}} to control gc directly.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:10)

out <- mrgsolve::mrgsim(mod, data)

obj <- as_mrgsim_ds(out)

obj

}
\seealso{
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
