% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{current_location}
\alias{current_location}
\title{Get the current location of mrgsimsds object files}
\usage{
current_location(x)
}
\arguments{
\item{x}{an mrgsimsds object.}
}
\description{
Get the current location of mrgsimsds object files
}
