% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_ds}
\alias{file_ds}
\title{Create an output file name}
\usage{
file_ds(id = NULL)
}
\arguments{
\item{id}{a tag used to form the file name; if not provided, a random name
will be generated.}
}
\value{
A character file name.
}
\description{
Create an output file name
}
\examples{
file_ds()

}
