% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{house_ds}
\alias{house_ds}
\title{Return a pre-compiled, PK/PD model for 'Apache' 'Arrow'-backed simulation
outputs}
\usage{
house_ds(...)
}
\arguments{
\item{...}{passed to \code{\link[mrgsolve:house]{mrgsolve::house()}}.}
}
\description{
This is a very-light wrapper around \code{\link[mrgsolve:house]{mrgsolve::house()}}.
}
\seealso{
\code{\link[=mcode_ds]{mcode_ds()}}, \code{\link[=mread_ds]{mread_ds()}}, \code{\link[=mcode_ds]{mcode_ds()}}.
}
