% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mcode_ds}
\alias{mcode_ds}
\title{Write, compile, and load model code for 'Apache' 'Arrow'-backed simulation
outputs}
\usage{
mcode_ds(...)
}
\arguments{
\item{...}{passed to \code{\link[mrgsolve:mcode]{mrgsolve::mcode()}}.}
}
\description{
This is a very-light wrapper around \code{\link[mrgsolve:mcode]{mrgsolve::mcode()}}.
}
\seealso{
\code{\link[=mread_ds]{mread_ds()}}, \code{\link[=modlib_ds]{modlib_ds()}}, \code{\link[=house_ds]{house_ds()}}.
}
