% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mread_ds}
\alias{mread_ds}
\alias{mcode_ds}
\alias{modlib_ds}
\alias{house_ds}
\alias{mread_cache_ds}
\title{Load an mrgsolve model for Arrow-backed simulation}
\usage{
mread_ds(...)

mcode_ds(...)

modlib_ds(...)

house_ds(...)

mread_cache_ds(...)
}
\arguments{
\item{...}{passed to the corresponding mrgsolve function. \code{sloc} is already
specified underneath and cannot be included here.}
}
\value{
A model object with process information saved, suitable for use with
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
\description{
Thin wrappers around mrgsolve model-loading functions (\code{mread()},
\code{mcode()}, \code{modlib()}, \code{house()}, \code{mread_cache()}) that additionally call
\code{\link[=save_process_info]{save_process_info()}} to stamp the model with the current process ID. This
stamp is required by \code{\link[=mrgsim_ds]{mrgsim_ds()}} to correctly associate simulation outputs
with the process that created them.
}
\examples{
mod <- house_ds()

mod

}
\seealso{
\code{\link[=save_process_info]{save_process_info()}}.
}
