% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mrgsimsds}
\alias{print.mrgsimsds}
\title{Print an mrgsimsds object}
\usage{
\method{print}{mrgsimsds}(x, n = 8, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{n}{number of rows to show from the cached head data.}

\item{...}{not used.}
}
\value{
\code{x} invisibly.
}
\description{
Print an mrgsimsds object
}
\examples{
mod <- house_ds(end = 24)

out <- mrgsim_ds(mod, events = ev(amt = 100))

print(out)

}
