% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_ds}
\alias{prune_ds}
\alias{prune_ds.mrgsimsds}
\alias{prune_ds.list}
\title{Prune a list of mrgsimsds objects}
\usage{
prune_ds(x, ..., inform = TRUE)

\method{prune_ds}{mrgsimsds}(x, ...)

\method{prune_ds}{list}(x, ..., inform = TRUE)
}
\arguments{
\item{x}{a list of R objects or a single mrgsimsds object.}

\item{...}{not used.}

\item{inform}{(list method only) issue a message when objects in some list
slots are dropped.}
}
\value{
When \code{x} is a list, it will be returned with only the mrgsimsds objects
retained. If no mrgsimsds objects are found, an empty list is returned with
a warning.

When \code{x} is an mrgsimsds object, it will be invisibly returned.
}
\description{
Filters a mixed list down to only the elements that are \code{mrgsimsds} objects,
dropping anything else (e.g. \code{NULL}, data frames, character vectors). When
passed a single \code{mrgsimsds} object it is returned invisibly unchanged.
}
\examples{
mod <- house_ds(end = 24)

out <- mrgsim_ds(mod, events = ev(amt = 100))

sims <- list(out, letters)

prune_ds(sims)

}
