% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_ds}
\alias{refresh_ds}
\alias{refresh_ds.mrgsimsds}
\alias{refresh_ds.list}
\title{Refresh 'Arrow' dataset pointers}
\usage{
refresh_ds(x, ...)

\method{refresh_ds}{mrgsimsds}(x, ...)

\method{refresh_ds}{list}(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{for future use.}
}
\value{
The mrgsimsds object is returned invisibly with pointers refreshed;
modification is made in place.
}
\description{
Pointers to arrow data sets will be invalid when the simulation is run in a
different process, for example when simulating in parallel. The pointers
should be refreshed on the head node once the simulation is finished.
}
\details{
To refresh the pointers, \code{refresh_ds()} checks that the files still exist
and passes the file list to \code{\link[arrow:open_dataset]{arrow::open_dataset()}}. The object \code{pid} and
the \code{dim} attributes are also refreshed, after re-opening the data set.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:100)

out <- lapply(1:3, function(rep) {
  mrgsim_ds(mod, data) 
})

refresh_ds(out)

}
