% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_output_ds}
\alias{set_output_ds}
\title{Set default output directory}
\usage{
set_output_ds(x, output_dir)
}
\arguments{
\item{x}{a model object.}

\item{output_dir}{path to output directory.}
}
\description{
Set default output directory
}
