% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{set_temp_ds}
\alias{set_temp_ds}
\alias{get_temp_ds}
\title{Set or get default temporary output directory}
\usage{
set_temp_ds(x)

get_temp_ds(x)
}
\arguments{
\item{x}{a model object.}
}
\description{
Set or get default temporary output directory
}
\examples{
mod <- mrgsolve::house()

mod <- set_temp_ds(mod)

get_temp_ds(mod)

}
