% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.R
\name{take_ownership}
\alias{take_ownership}
\alias{check_ownership}
\alias{list_ownership}
\alias{ownership}
\alias{disown}
\title{Take or check ownership of simulation files}
\usage{
take_ownership(x)

check_ownership(x)

list_ownership(full.names = FALSE)

ownership()

disown(x)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{full.names}{if \code{TRUE}, include the directory path when listing file
ownership.}
}
\description{
Take or check ownership of simulation files
}
