% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_arrow_ds}
\alias{as_arrow_ds}
\title{Coerce an mrgsimsds object to an arrow data set}
\usage{
as_arrow_ds(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object or a list of mrgsimsds objects.}

\item{...}{not used.}
}
\value{
An 'Apache' 'Arrow' \link[arrow:Dataset]{arrow::Dataset} object.
}
\description{
Coerce an mrgsimsds object to an arrow data set
}
\details{
The method for list will retain only list positions containing an \code{mrgsimsds}
object. A single data set object is returned.
}
\examples{
mod <- house_ds(end = 5)

out <- mrgsim_ds(mod, events = ev(amt = 100))

as_arrow_ds(out)

}
