% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_arrow_ds}
\alias{as_arrow_ds}
\title{Coerce an mrgsimsds object to an arrow data set}
\usage{
as_arrow_ds(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{not used.}
}
\description{
Coerce an mrgsimsds object to an arrow data set
}
