% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_arrow_table.mrgsimsds}
\alias{as_arrow_table.mrgsimsds}
\title{Coerce an mrgsimsds object to an arrow table}
\usage{
\method{as_arrow_table}{mrgsimsds}(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{passed to \code{\link[arrow:as_arrow_table]{arrow::as_arrow_table()}}.}
}
\description{
Coerce an mrgsimsds object to an arrow table
}
