% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_duckdb_ds}
\alias{as_duckdb_ds}
\title{Coerce an mrgsimsds object to a DuckDB table}
\usage{
as_duckdb_ds(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object or a list of mrgsimsds objects.}

\item{...}{passed to \code{\link[=as_arrow_ds]{as_arrow_ds()}}.}
}
\description{
Coerce an mrgsimsds object to a DuckDB table
}
\details{
The conversaion is handled by \code{\link[=as_arrow_ds]{as_arrow_ds()}}.
}
\seealso{
\code{\link[=as_arrow_ds]{as_arrow_ds()}}
}
