% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{as_mrgsim_ds}
\alias{as_mrgsim_ds}
\title{Coerce an mrgsims object to arrow-backed mrgsimsds object}
\usage{
as_mrgsim_ds(x, file = temp_file(x), verbose = FALSE)
}
\arguments{
\item{x}{an mrgsims object.}

\item{file}{path to file where output will be written using
\code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Coerce an mrgsims object to arrow-backed mrgsimsds object
}
\examples{
mod <- mrgsolve::house()
data <- mrgsolve::ev_expand(amt = 100, ID = 1:10)
out <- mrgsolve::mrgsim(mod, data)
obj <- as_mrgsim_ds(out)

}
\seealso{
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
