% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{as_mrgsim_ds}
\alias{as_mrgsim_ds}
\title{Coerce an mrgsims object to 'Arrow'-backed mrgsimsds object}
\usage{
as_mrgsim_ds(x, id = NULL, verbose = FALSE, gc = TRUE)
}
\arguments{
\item{x}{an mrgsims object.}

\item{id}{used to label output files; see details.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{gc}{if \code{TRUE}, a finalizer function will attempt to remove files once
the object is out of scope.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Coerce an mrgsims object to 'Arrow'-backed mrgsimsds object
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:10)

out <- mrgsolve::mrgsim(mod, data)

obj <- as_mrgsim_ds(out)

obj

}
\seealso{
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
