% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-ds.R
\name{as_mrgsim_ds}
\alias{as_mrgsim_ds}
\title{Coerce an mrgsims object to 'Apache' 'Arrow'-backed mrgsimsds object}
\usage{
as_mrgsim_ds(x, verbose = FALSE, gc = TRUE)
}
\arguments{
\item{x}{an mrgsims object.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{gc}{if \code{TRUE}, a finalizer function will attempt to remove files once
the object is out of scope; set to \code{FALSE} to protect from automatic
cleanup, or use \code{\link[=move_ds]{move_ds()}} or \code{\link[=write_ds]{write_ds()}} to relocate files outside of
\code{tempdir()}.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Converts the output of \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}} to an \code{mrgsimsds} object by
writing the simulation data to a parquet file in \code{tempdir()}. See
\code{\link[=move_ds]{move_ds()}} or \code{\link[=write_ds]{write_ds()}} to change the location of the files, protecting
them from the garbage collector.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:10)

out <- mrgsolve::mrgsim(mod, data)

obj <- as_mrgsim_ds(out)

obj

}
\seealso{
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
