% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_tibble.mrgsimsds}
\alias{as_tibble.mrgsimsds}
\alias{collect.mrgsimsds}
\title{Coerce an mrgsimsds object to a tibble}
\usage{
\method{as_tibble}{mrgsimsds}(x, ...)

\method{collect}{mrgsimsds}(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{passed to \code{\link[dplyr:reexports]{dplyr::as_tibble()}} or
\code{\link[dplyr:compute]{dplyr::collect()}}.}
}
\description{
Coerce an mrgsimsds object to a tibble
}
