% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_tibble.mrgsimsds}
\alias{as_tibble.mrgsimsds}
\alias{collect.mrgsimsds}
\alias{as.data.frame.mrgsimsds}
\title{Coerce an mrgsimsds object to a tbl}
\usage{
\method{as_tibble}{mrgsimsds}(x, ...)

\method{collect}{mrgsimsds}(x, ...)

\method{as.data.frame}{mrgsimsds}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{passed to \code{\link[dplyr:reexports]{dplyr::as_tibble()}} or \code{\link[dplyr:compute]{dplyr::collect()}}.}

\item{row.names}{passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}.}

\item{optional}{passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}.}
}
\value{
A \code{tbl} containing simulated data.
}
\description{
Coerce an mrgsimsds object to a tbl
}
\examples{
mod <- house_ds(end = 5)

out <- mrgsim_ds(mod, events = ev(amt = 100))

as.data.frame(out)

tibble::as_tibble(out)

dplyr::collect(out)

}
