% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.R
\name{copy_ds}
\alias{copy_ds}
\title{Copy an mrgsims object}
\usage{
copy_ds(x, own = FALSE)
}
\arguments{
\item{x}{the object to copy.}
}
\value{
An mrgsims object with identical fields, but updated pid.
}
\description{
Copy an mrgsims object
}
