% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc.R
\name{gc_ds}
\alias{gc_ds}
\title{Set collection status for mrgsimsds objects}
\usage{
gc_ds(x, ...)
}
\arguments{
\item{x}{a list of mrgsimsds objects or a single mrgsimsds object.}

\item{...}{not used.}

\item{value}{logical; if \code{TRUE} the underlying files will be cleaned up.}

\item{notify}{logical; if \code{TRUE} a message will be issued when files are
cleaned up; this is intended for debugging or troubleshooting purposes.}
}
\value{
An mrgsimsds object or a list of those objects is returned, potentially
with the \code{gc} status updated.
}
\description{
Set collection status for mrgsimsds objects
}
\examples{
mod <- modlib_ds("popex", outvars = "IPRED")

data <- ev_expand(amt = 100, ID = 1:5)

out <- mrgsim_ds(mod, data)

out <- gc_ds(out, value = FALSE)

out <- gc_ds(out, value = TRUE)

out <- lapply(1:3, function(rep) {
  out <- mrgsim_ds(mod, data) 
  out
})

out <- gc_ds(out, value = FALSE)

}
