% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{list_temp}
\alias{list_temp}
\alias{retain_temp}
\alias{purge_temp}
\title{Manage simulated outputs in the per-session temporary directory}
\usage{
list_temp()

retain_temp(...)

purge_temp()
}
\arguments{
\item{...}{objects whose files will not be purged.}
}
\description{
Manage simulated outputs in the per-session temporary directory
}
\examples{
mod <- house_ds()

out <- lapply(1:10, \(x) mrgsim_ds(mod))

list_temp()

sims <- reduce_ds(out)

list_temp()

retain_temp(sims)

list_temp() 

purge_temp() 

list_temp()

}
