% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{list_temp}
\alias{list_temp}
\alias{retain_temp}
\alias{purge_temp}
\title{Manage simulated outputs in the per-session temporary directory}
\usage{
list_temp(quietly = FALSE)

retain_temp(..., quietly = FALSE)

purge_temp(quietly = FALSE)
}
\arguments{
\item{quietly}{if \code{TRUE}, messages will be suppressed.}

\item{...}{objects whose files will be retained.}
}
\value{
\code{list_temp()} returns a vector of file names. Other functions return \code{NULL}.
All return values are invisible.
}
\description{
Note: \code{purge_temp()} should not be needed or used in routine usage when
simulation output objects are subject to the garbage collector. If you
run \code{purge_temp()} with active objects pointing to those files, you will
get an error when trying to access the data.
}
\examples{
mod <- house_ds()

out <- lapply(1:10, \(x) mrgsim_ds(mod))

list_temp()

sims <- reduce_ds(out)

list_temp()

retain_temp(sims)

list_temp() 

purge_temp() 

list_temp()

}
