% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{modlib_ds}
\alias{modlib_ds}
\title{Internal model library for 'Arrow'-backed simulation outputs}
\usage{
modlib_ds(..., output_dir = NULL)
}
\arguments{
\item{...}{passed to \code{\link[mrgsolve:modlib]{mrgsolve::modlib()}}.}
}
\description{
This is a very-light wrapper around \code{\link[mrgsolve:modlib]{mrgsolve::modlib()}}.
}
\details{
After calling \code{modlib()}, the model object is passed through either
\code{\link[=set_output_ds]{set_output_ds()}} or \code{\link[=set_temp_ds]{set_temp_ds()}}.
}
\seealso{
\code{\link[=mread_ds]{mread_ds()}}, \code{\link[=mcode_ds]{mcode_ds()}}, \code{\link[=house_ds]{house_ds()}}.
}
