% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_ds}
\alias{move_ds}
\alias{write_ds}
\title{Move data set files to a new directory.}
\usage{
move_ds(x, path)

write_ds(x, sink, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{path}{the new directory location for backing files.}

\item{sink}{the complete path (including file name) for a single parquet
file containing all simulated data.}

\item{...}{passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}
}
\value{
Both functions return the mrgsimsds object; it is critical to capture the
return value in order to continue working with the object in the current
R session.
}
\description{
Use \code{move_ds()} to just change the enclosing directory. \code{write_ds()} can also
move the files, but also condenses all simulation output in to a single
parquet file if multiple files are backing the mrgsimsds object.
}
\examples{

mod <- house_ds()

out <- mrgsim_ds(mod, events = ev(amt = 100))

out <- write_ds(out, sink = file.path(tempdir(), "example.parquet"))

out$files

\dontrun{
  out <- move_ds(out, path = "data/simulated") 
}

}
