% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_ds}
\alias{move_ds}
\alias{rename_ds}
\alias{write_ds}
\title{Move data set files to a new directory.}
\usage{
move_ds(x, path)

rename_ds(x, id)

write_ds(x, sink, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{path}{the new directory location for backing files.}

\item{id}{a short name used to create data set files for the simulated
output.}

\item{sink}{the complete path (including file name) for a single parquet
file containing all simulated data.}

\item{...}{passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}
}
\value{
Both functions return the mrgsimsds object; it is critical to capture the
return value in order to continue working with the object in the current
R session.
}
\description{
Use \code{move_ds()} to just change the enclosing directory. \code{write_ds()} can also
move the files, but also condenses all simulation output in to a single
parquet file if multiple files are backing the mrgsimsds object. See
\emph{Details}.
}
\details{
When dataset files are rewritten to a single file with \code{write_ds()}, those
files will no longer be cleaned up when the containing R object is finalized
upon garbage collection. When dataset files are moved outside of \code{tempdir()},
those files, too, will no longer be cleaned up on garbage collection; but
file cleanup will continue to occur as long as the files remain under
\code{tempdir()}. No change in finalization behavior due to garbage collection
of the containing object will happen when files are renamed.
}
\examples{

mod <- house_ds()

out <- mrgsim_ds(mod, events = ev(amt = 100))

out <- write_ds(out, sink = file.path(tempdir(), "example.parquet"))

out$files

\dontrun{
  out <- move_ds(out, path = "data/simulated") 
}

}
