% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_ds}
\alias{move_ds}
\alias{rename_ds}
\alias{write_ds}
\title{Move, rename, or write out data set files}
\usage{
move_ds(x, path)

rename_ds(x, id)

write_ds(x, sink, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{path}{the new directory location for backing files.}

\item{id}{a short name used to create data set files for the simulated
output.}

\item{sink}{the complete path (including file name) for a single parquet
file containing all simulated data; passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}

\item{...}{passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}; files are always written
in parquet format.}
}
\value{
All three functions return the new file list, invisibly.
}
\description{
Use \code{move_ds()} to change the enclosing directory. \code{write_ds()} can also
move the files, but it always condenses th simulation output into a single
parquet file if multiple files are backing the mrgsimsds object. All
operations are made on the object in place; see \strong{Details}.
}
\details{
There is an important distinction between \code{write_ds()} and \code{move_ds()} or
\code{rename_ds()} for multi-file objects. The backing files can be moved or
written easily, without much computational effort. For multi-file simulation
outputs, \code{write_ds()} will need to read each file and then write the data
out to a single file. Apache Arrow can do this very efficiently, but there
will still be an additional, potentially noticeable computational effort

When dataset files are rewritten to a single file with \code{write_ds()}, those
files will no longer be cleaned up when the containing R object is finalized
upon garbage collection. When dataset files are moved outside of \code{tempdir()},
those files, too, will no longer be cleaned up on garbage collection; but
file cleanup will continue to occur as long as the files remain under
\code{tempdir()}. No change in finalization behavior due to garbage collection
of the containing object will happen when files are renamed.

The object (\code{x}) is required to own the underlying files in order to move
them with \code{move_ds()}; ownership is not required for \code{write_ds()}.

All three functions modify \code{x} in place and file ownership stays with \code{x}.
}
\examples{

mod <- house_ds()

out <- lapply(1:3, \(x) { mrgsim_ds(mod, events = ev(amt = 100)) })

out <- reduce_ds(out)

rename_ds(out, id = "example-sims")

basename(out$files)

write_ds(out, sink = file.path(tempdir(), "example.parquet"))

basename(out$files)

\dontrun{
  move_ds(out, path = "data/simulated") 
}

}
