% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mread_ds}
\alias{mread_ds}
\title{Read a model specification file for 'Apache' 'Arrow'-backed simulation
outputs}
\usage{
mread_ds(...)
}
\arguments{
\item{...}{passed to \code{\link[mrgsolve:mread]{mrgsolve::mread()}}.}
}
\description{
This is a very-light wrapper around \code{\link[mrgsolve:mread]{mrgsolve::mread()}}.
}
\seealso{
\code{\link[=mcode_ds]{mcode_ds()}}, \code{\link[=modlib_ds]{modlib_ds()}}, \code{\link[=house_ds]{house_ds()}}.
}
