% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mread_ds}
\alias{mread_ds}
\title{Read a model specification file for 'Apache' 'Arrow'-backed simulation
outputs}
\usage{
mread_ds(..., output_dir = NULL)
}
\arguments{
\item{...}{passed to \code{\link[mrgsolve:mread]{mrgsolve::mread()}}.}
}
\description{
This is a very-light wrapper around \code{\link[mrgsolve:mread]{mrgsolve::mread()}}.
}
\details{
After calling \code{mcode()}, the model object is passed through either
\code{\link[=set_output_ds]{set_output_ds()}} or \code{\link[=set_temp_ds]{set_temp_ds()}}.
}
\seealso{
\code{\link[=mcode_ds]{mcode_ds()}}, \code{\link[=modlib_ds]{modlib_ds()}}, \code{\link[=house_ds]{house_ds()}}.
}
