% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mread_ds}
\alias{mread_ds}
\alias{mcode_ds}
\alias{modlib_ds}
\alias{house_ds}
\alias{mread_cache_ds}
\title{Read a model specification file for 'Apache' 'Arrow'-backed simulation
outputs}
\usage{
mread_ds(...)

mcode_ds(...)

modlib_ds(...)

house_ds(...)

mread_cache_ds(...)
}
\arguments{
\item{...}{passed to the corresponding mrgsolve function.}
}
\value{
A model object with process information saved, suitable for use with
\code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
\description{
These are very-light wrappers around mrgsolve functions used to load
models for simulation.
}
\examples{
mod <- house_ds()

mod

}
\seealso{
\code{\link[=save_process_info]{save_process_info()}}.
}
