% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{mrgsim.ds-package}
\alias{mrgsim.ds-package}
\title{mrgsim.ds: 'Apache' 'Arrow' Dataset-Backed Simulation Outputs for 'mrgsolve'}
\description{
Provides tools for creating and managing file streams in support of large simulation or other outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kylebaron/mrgsim.ds}
  \item \url{https://kylebaron.github.io/mrgsim.ds/}
  \item Report bugs at \url{https://github.com/kylebaron/mrgim.ds/issues}
}

}
\author{
\strong{Maintainer}: Kyle T Baron \email{kylebtwin@imap.cc} (\href{https://orcid.org/0000-0001-7252-5656}{ORCID}) [copyright holder]

}
\keyword{internal}
