% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{mrgsim.ds}
\alias{mrgsim.ds-package}
\alias{mrgsim.ds}
\title{mrgsim.ds: 'Apache' 'Arrow' Dataset-Backed Simulation Outputs for 'mrgsolve'}
\description{
\code{mrgsim.ds} provides an \href{https://arrow.apache.org/docs/r/}{Apache Arrow}-backed
simulation output object for \href{https://mrgsolve.org}{mrgsolve}, greatly reducing
the memory footprint of large simulations and providing a high-performance
pipeline for summarizing huge simulation outputs. The arrow-based simulation
output objects in R claim ownership of their files on disk.
Those files are automatically removed when the owning object goes out of scope
and becomes subject to the R garbage collector. While "anonymous",
parquet-formatted files hold the data in \code{tempdir()} as you are working in
R, functions are provided to move this data to more permanent locations for
later use.
}
\section{Function listing}{

\itemize{
\item Load models
\itemize{
\item \code{\link[=mread_ds]{mread_ds()}}
\item \code{\link[=mcode_ds]{mcode_ds()}}
\item \code{\link[=mread_cache_ds]{mread_cache_ds()}}
\item \code{\link[=modlib_ds]{modlib_ds()}}
\item \code{\link[=house_ds]{house_ds()}}
}
\item Generate Apache Arrow dataset-backed outputs
\itemize{
\item \code{\link[=mrgsim_ds]{mrgsim_ds()}}
\item \code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}
}
\item S3 Methods
\itemize{
\item \code{\link[=head.mrgsimsds]{head.mrgsimsds()}}
\item \code{\link[=tail.mrgsimsds]{tail.mrgsimsds()}}
\item \code{\link[=dim.mrgsimsds]{dim.mrgsimsds()}}
\item \code{\link[=names.mrgsimsds]{names.mrgsimsds()}}
}
\item Move, rename, or combine files
\itemize{
\item \code{\link[=move_ds]{move_ds()}}
\item \code{\link[=rename_ds]{rename_ds()}}
\item \code{\link[=combine_ds]{combine_ds()}}
}
\item Save and restore
\itemize{
\item \code{\link[=save_ds]{save_ds()}}
\item \code{\link[=read_ds]{read_ds()}}
\item \code{\link[=write_parquet_ds]{write_parquet_ds()}}
\item \code{\link[=write_dataset_ds]{write_dataset_ds()}}
}
\item Ownership
\itemize{
\item \code{\link[=ownership]{ownership()}}
\item \code{\link[=check_ownership]{check_ownership()}}
\item \code{\link[=list_ownership]{list_ownership()}}
\item \code{\link[=take_ownership]{take_ownership()}}
\item \code{\link[=disown]{disown()}}
\item \code{\link[=copy_ds]{copy_ds()}}
}
\item Work with lists of outputs
\itemize{
\item \code{\link[=reduce_ds]{reduce_ds()}}
\item \code{\link[=refresh_ds]{refresh_ds()}}
\item \code{\link[=prune_ds]{prune_ds()}}
}
\item Manage tempdir
\itemize{
\item \code{\link[=list_temp]{list_temp()}}
\item \code{\link[=purge_temp]{purge_temp()}}
}
\item Enter dplyr / arrow pipelines with
\itemize{
\item \code{\link[dplyr:mutate]{dplyr::mutate()}}
\item \code{\link[dplyr:select]{dplyr::select()}}
\item \code{\link[dplyr:filter]{dplyr::filter()}}
\item \code{\link[dplyr:summarise]{dplyr::summarise()}}
\item \code{\link[dplyr:summarise]{dplyr::summarize()}}
\item \code{\link[dplyr:rename]{dplyr::rename()}}
\item \code{\link[dplyr:arrange]{dplyr::arrange()}}
\item \code{\link[dplyr:group_by]{dplyr::group_by()}}
\item \code{\link[dplyr:distinct]{dplyr::distinct()}}
\item \code{\link[dplyr:relocate]{dplyr::relocate()}}
\item \code{\link[dplyr:count]{dplyr::count()}}
\item \code{\link[dplyr:pull]{dplyr::pull()}}
}
\item Coerce to R objects
\itemize{
\item \code{\link[=as.data.frame]{as.data.frame()}}
\item \code{\link[dplyr:reexports]{dplyr::as_tibble()}}
\item \code{\link[dplyr:compute]{dplyr::collect()}}
\item \code{\link[=as_arrow_ds]{as_arrow_ds()}}
\item \code{\link[arrow:as_arrow_table]{arrow::as_arrow_table()}}
\item \code{\link[=as_duckdb_ds]{as_duckdb_ds()}}
}
}
}

\examples{

mod <- house_ds(end = 32)

data <- evd_expand(amt = seq(100, 300, 10))

out <- mrgsim_ds(mod, data)

out

head(out)

tail(out)

plot(out, nid = 10)

list_temp()

ownership()

\dontrun{

rename_ds(out, "reg-100-300")

list_temp()

move_ds(out, "data/sim/regimens")
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://metrumresearchgroup.github.io/mrgsim.ds/}
  \item \url{https://github.com/metrumresearchgroup/mrgsim.ds}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/mrgsim.ds/issues}
}

}
\author{
\strong{Maintainer}: Kyle T Baron \email{kyleb@metrumrg.com} (\href{https://orcid.org/0000-0001-7252-5656}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kyle Meyer
}

Other contributors:
\itemize{
  \item Metrum Research Group [copyright holder, funder]
}

}
