% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{mrgsim_ds}
\alias{mrgsim_ds}
\title{Simulate from a model object, returning an arrow-backed output object}
\usage{
mrgsim_ds(x, ..., id = NULL, tags = list(), verbose = FALSE)
}
\arguments{
\item{x}{a model object.}

\item{...}{passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}}.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{tag}{a named list of atomic data to tag (or mutate) the simulated
output.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Simulate from a model object, returning an arrow-backed output object
}
\examples{
mod <- mrgsolve::house()
data <- mrgsolve::ev_expand(amt = 100, ID = 1:10)
out <- mrgsim_ds(mod, data, end = 72, delta = 0.1)

out <- mrgsim_ds(mod, data, tag = list(rep = 1))
head(out)

}
