% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-ds.R
\name{mrgsim_ds}
\alias{mrgsim_ds}
\title{Simulate from a model object, returning an 'Apache' 'Arrow'-backed output object}
\usage{
mrgsim_ds(x, ..., tags = list(), verbose = FALSE, gc = TRUE)
}
\arguments{
\item{x}{a model object loaded through \code{\link[=mread_ds]{mread_ds()}}, \code{\link[=mcode_ds]{mcode_ds()}},
\code{\link[=modlib_ds]{modlib_ds()}}, \code{\link[=mread_cache_ds]{mread_cache_ds()}}, or \code{\link[=house_ds]{house_ds()}}.}

\item{...}{passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}}.}

\item{tags}{a named list of atomic data to tag (or mutate) the simulated
output.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{gc}{initial gc setting; if \code{TRUE}, a finalizer function will attempt
to remove files once the object is out of scope. This value is not locked:
\code{\link[=move_ds]{move_ds()}} and \code{\link[=save_ds]{save_ds()}} will automatically adjust gc based on whether
the files remain under the temporary directory. To lock the gc setting and
prevent automatic adjustment, call \code{\link[=gc_ds]{gc_ds()}} after creation.}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
Runs \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}} and writes simulation output to a parquet file in
a temporary directory, returning an \code{mrgsimsds} object. Files in the
temporary directory are auto-deleted on garbage collection by default. Use
\code{\link[=move_ds]{move_ds()}} or \code{\link[=save_ds]{save_ds()}} to relocate files outside of the temporary
directory, which automatically disables gc, or call \code{\link[=gc_ds]{gc_ds()}} to control gc
directly. Note that full argument names must be used for all arguments.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:10)

out <- mrgsim_ds(mod, data, end = 72, delta = 0.1)

out <- mrgsim_ds(mod, data, tags = list(rep = 1))

head(out)

}
\seealso{
\code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}, \link{mrgsimsds-methods}.
}
