% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-ds.R
\name{mrgsim_ds}
\alias{mrgsim_ds}
\title{Simulate from a model object, returning an arrow-backed output object}
\usage{
mrgsim_ds(x, ..., tags = list(), verbose = FALSE, gc = TRUE)
}
\arguments{
\item{x}{a model object loaded through \code{\link[=mread_ds]{mread_ds()}}, \code{\link[=mcode_ds]{mcode_ds()}},
\code{\link[=modlib_ds]{modlib_ds()}}, \code{\link[=mread_cache_ds]{mread_cache_ds()}}, or \code{\link[=house_ds]{house_ds()}}.}

\item{...}{passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}}.}

\item{tags}{a named list of atomic data to tag (or mutate) the simulated
output.}

\item{verbose}{if \code{TRUE}, print progress information to the console.}

\item{gc}{if \code{TRUE}, a finalizer function will attempt to remove files once
the object is out of scope; set to \code{FALSE} to protect from automatic
cleanup. Otherwise, move the files from \code{tempdir()}}
}
\value{
An object with class \code{mrgsimsds}.
}
\description{
All simulation data is saved to \code{tempdir()} according to the parent or head
node that the computation is run from. See \code{\link[=move_ds]{move_ds()}} to \code{\link[=write_ds]{write_ds()}}
change the location of the files, protecting them from the garbage collector.
Note that full names must be used for all arguments.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:10)

out <- mrgsim_ds(mod, data, end = 72, delta = 0.1)

out <- mrgsim_ds(mod, data, tags = list(rep = 1))

head(out)

}
\seealso{
\code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}, \link{mrgsimsds-methods}.
}
