% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{mrgsimsds-methods}
\alias{mrgsimsds-methods}
\alias{dim.mrgsimsds}
\alias{nrow.mrgsimsds}
\alias{ncol.mrgsimsds}
\alias{head.mrgsimsds}
\alias{tail.mrgsimsds}
\alias{plot.mrgsimsds}
\title{Interact with mrgsimsds objects}
\usage{
\method{dim}{mrgsimsds}(x)

nrow.mrgsimsds(x)

ncol.mrgsimsds(x)

\method{head}{mrgsimsds}(x, n = 6L, ...)

\method{tail}{mrgsimsds}(x, ...)

\method{plot}{mrgsimsds}(
  x,
  y = NULL,
  ...,
  nid = 5,
  batch_size = 20000,
  logy = FALSE,
  .dots = list()
)
}
\arguments{
\item{x}{an mrgsimsds object, output from
\code{\link[=mrgsim_ds]{mrgsim_ds()}} or \code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}.}

\item{n}{number of rows to return.}

\item{...}{arguments to be passed to or from other methods.}

\item{y}{not used.}
}
\description{
Interact with mrgsimsds objects
}
\examples{
mod <- mrgsolve::house()
out <- mrgsim_ds(mod, events = mrgsolve::ev(amt = 100))

dim(out)
head(out)
nrow(out)
ncol(out)
head(out)
try(tail(out))

}
