% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{mrgsimsds-methods}
\alias{mrgsimsds-methods}
\alias{dim.mrgsimsds}
\alias{head.mrgsimsds}
\alias{tail.mrgsimsds}
\alias{plot.mrgsimsds}
\title{Interact with mrgsimsds objects}
\usage{
\method{dim}{mrgsimsds}(x)

\method{head}{mrgsimsds}(x, n = 6L, ...)

\method{tail}{mrgsimsds}(x, n = 6L, ...)

\method{plot}{mrgsimsds}(
  x,
  y = NULL,
  ...,
  nid = 5,
  batch_size = 20000,
  logy = FALSE,
  .dots = list()
)
}
\arguments{
\item{x}{an mrgsimsds object, output from
\code{\link[=mrgsim_ds]{mrgsim_ds()}} or \code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}.}

\item{n}{number of rows to return.}

\item{...}{arguments to be passed to or from other methods.}

\item{y}{not used.}
}
\description{
Interact with mrgsimsds objects
}
\details{
\code{head()} and \code{tail()} only look at the first and last file in the data
set, respectively.
}
\examples{
mod <- house_ds()

out <- mrgsim_ds(mod, events = ev(amt = 100))

\dontrun{
dim(out)
head(out)
tail(out)
nrow(out)
ncol(out)
plot(out)
}

}
