% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsimsds.R
\name{mrgsimsds-methods}
\alias{mrgsimsds-methods}
\alias{dim.mrgsimsds}
\alias{head.mrgsimsds}
\alias{tail.mrgsimsds}
\alias{plot.mrgsimsds}
\title{Interact with mrgsimsds objects}
\usage{
\method{dim}{mrgsimsds}(x)

\method{head}{mrgsimsds}(x, n = 6L, ...)

\method{tail}{mrgsimsds}(x, n = 6L, ...)

\method{plot}{mrgsimsds}(
  x,
  y = NULL,
  ...,
  nid = 5,
  batch_size = 20000,
  logy = FALSE,
  .dots = list()
)
}
\arguments{
\item{x}{an mrgsimsds object, output from
\code{\link[=mrgsim_ds]{mrgsim_ds()}} or \code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}.}

\item{n}{number of rows to return.}

\item{...}{arguments to be passed to or from other methods.}

\item{y}{a formula for plotting simulated data; if not provided, all
columns will be plotted.}
}
\description{
Interact with mrgsimsds objects
}
\details{
\code{head()} and \code{tail()} only look at the first and last file in the data
set, respectively.
}
\examples{
mod <- house_ds(end = 24)

mod <- omat(mod, diag(0.04, 4))

data <- ev_expand(amt = c(100, 300), ID = 1:20)

set.seed(10203)

out <- mrgsim_ds(mod, data = data)

dim(out)
head(out)
tail(out)
nrow(out)
ncol(out)
plot(out, ~ CP + RESP, nid = 10)

}
