% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-ds.R
\name{mrgsimsds-methods}
\alias{mrgsimsds-methods}
\alias{dim.mrgsimsds}
\alias{head.mrgsimsds}
\alias{tail.mrgsimsds}
\alias{names.mrgsimsds}
\alias{plot.mrgsimsds}
\title{Interact with mrgsimsds objects}
\usage{
\method{dim}{mrgsimsds}(x)

\method{head}{mrgsimsds}(x, n = 6L, ...)

\method{tail}{mrgsimsds}(x, n = 6L, ...)

\method{names}{mrgsimsds}(x)

\method{plot}{mrgsimsds}(
  x,
  y = NULL,
  ...,
  nid = 16,
  batch_size = 20000,
  logy = FALSE,
  .dots = list()
)
}
\arguments{
\item{x}{an mrgsimsds object, output from
\code{\link[=mrgsim_ds]{mrgsim_ds()}} or \code{\link[=as_mrgsim_ds]{as_mrgsim_ds()}}.}

\item{n}{number of rows to return.}

\item{...}{arguments to be passed to or from other methods.}

\item{y}{a formula for plotting simulated data; if not provided, all
columns will be plotted.}

\item{nid}{number of subjects to plot.}

\item{batch_size}{size of batch when reading data for plot method.}

\item{logy}{if \code{TRUE}, plot data with log y-axis.}

\item{.dots}{a list of items to pass to \code{\link[mrgsolve:plot_sims]{mrgsolve::plot_sims()}}.}
}
\value{
\itemize{
\item \code{dim()}: integer vector of length 2 (rows, cols).
\item \code{head()}, \code{tail()}: a tibble of the first or last \code{n} rows.
\item \code{names()}: character vector of column names.
\item \code{plot()}: a plot object, returned invisibly.
}
}
\description{
Basic S3 methods for inspecting and plotting mrgsimsds objects: \code{dim()},
\code{head()}, \code{tail()}, \code{names()}, and \code{plot()}.
}
\details{
\code{head()}  and \code{tail()} only look at the first and last \code{file} in the data
set, respectively, when simulations are stored across multiple files. It is
possible this won't correspond to the first and last chunks rows of data
you will see when collecting the data via \code{\link[dplyr:compute]{dplyr::collect()}}.
}
\examples{
mod <- house_ds(end = 24)

mod <- omat(mod, diag(0.04, 4))

data <- ev_expand(amt = c(100, 300), ID = 1:20)

set.seed(10203)

out <- mrgsim_ds(mod, data = data)

dim(out)
head(out)
tail(out)
nrow(out)
ncol(out)
plot(out, ~ CP + RESP, nid = 10)

}
