% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{prune_ds}
\alias{prune_ds}
\title{Prune a list of mrgsimsds objects}
\usage{
prune_ds(x, ...)
}
\arguments{
\item{x}{a list of R objects or a single mrgsimsds object.}

\item{...}{not used.}

\item{inform}{issue a message when objects in some list slots are dropped.}
}
\value{
The input list of objects, with only the mrgsimsds objects retained.
}
\description{
Prune a list of mrgsimsds objects
}
\examples{
mod <- house_ds(end = 24)

out <- mrgsim_ds(mod, events = ev(amt = 100))

sims <- list(out, letters)

prune_ds(sims)

}
