% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce_ds}
\alias{reduce_ds}
\title{Reduce a list of mrgsimsds objects into a single object}
\usage{
reduce_ds(x, ...)
}
\arguments{
\item{x}{a list of mrgsimsds objects or a single mrgsimsds object.}

\item{...}{not used.}
}
\description{
Reduce a list of mrgsimsds objects into a single object
}
\details{
When \code{x} is a list, a new object is created and returned. This new object
will take ownership for all the files from the objects in the list.

When \code{x} is an mrgsimsds object, it will be returned invisibly with no
modification.
}
\examples{
mod <- modlib_ds("1005", outvars = "IPRED")

data <- ev_expand(amt = 100, ID = 1:100)

out <- lapply(1:3, function(rep) {
  out <- mrgsim_ds(mod, data) 
  out
})

length(out)

sims <- reduce_ds(out)

sims

check_ownership(sims)

lapply(out, check_ownership)

}
