% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_ds}
\alias{refresh_ds}
\alias{refresh_ds.mrgsimsds}
\alias{refresh_ds.list}
\title{Refresh 'Arrow' dataset pointers.}
\usage{
refresh_ds(x, ...)

\method{refresh_ds}{mrgsimsds}(x, ...)

\method{refresh_ds}{list}(x, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{...}{for future use.}
}
\value{
The mrgsimsds object is returned with pointers refreshed.
}
\description{
Pointers to arrow data sets will be invalid when the simulation is run in a
different process, for example when simulating in parallel. The pointers
should be refreshed on the head node once the simulation is finished.
}
\examples{
mod <- house_ds()

data <- ev_expand(amt = 100, ID = 1:100)

out <- lapply(1:3, function(rep) {
  out <- mrgsim_ds(mod, data) 
  out
})

out <- refresh_ds(out)

}
