% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_process_info}
\alias{save_process_info}
\title{Save information about the R process that loaded a model}
\usage{
save_process_info(x)
}
\arguments{
\item{x}{a model object.}
}
\value{
An updated model object suitable for using with \code{\link[=mrgsim_ds]{mrgsim_ds()}}.
}
\description{
Save information about the R process that loaded a model
}
\examples{
mod <- mrgsolve::house()

mod <- save_process_info(mod)

}
