% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{set_ds_dir}
\alias{set_ds_dir}
\alias{get_ds_dir}
\title{Set or get the output directory on a model object}
\usage{
set_ds_dir(mod, dir)

get_ds_dir(mod)
}
\arguments{
\item{mod}{a model object loaded with \code{\link[=mread_ds]{mread_ds()}} or equivalent.}

\item{dir}{the directory where simulated output should be written.}
}
\value{
\code{set_ds_dir()} returns the updated model object.

\code{get_ds_dir()} returns the output directory as a string.
}
\description{
Simulated output is written to a directory which is resolved one time, when
the model is loaded (see \code{\link[=mread_ds]{mread_ds()}}), and stamped on the model object. Use
\code{set_ds_dir()} to re-target a model after it was loaded and \code{get_ds_dir()}
to see where simulated output will be written.

The output directory is resolved in this order:
\enumerate{
\item the \code{dir} argument to \code{\link[=mread_ds]{mread_ds()}} and friends
\item \code{getOption("mrgsim.ds.dir")}
\item \code{tempdir()}
}

Setting the \code{mrgsim.ds.dir} option before loading a model is the most
convenient way to send output to a location which every worker can reach
when simulating in parallel; \code{tempdir()} on a worker node is generally
\emph{not} the same directory as \code{tempdir()} on the parent node.

The directory is created when it doesn't exist. Note that output files are
subject to garbage collection in the output directory whether or not that
directory is under \code{tempdir()}; see \code{\link[=move_ds]{move_ds()}} and \code{\link[=gc_ds]{gc_ds()}}.

Like other updates to the model environment, \code{set_ds_dir()} works by
reference: copies of a model object share one environment, so they will all
write to the directory you set. Load the model a second time (see
\code{\link[=mread_ds]{mread_ds()}}) when you want two model objects writing to two different
directories.
}
\examples{
mod <- house_ds()

get_ds_dir(mod)

mod <- set_ds_dir(mod, file.path(tempdir(), "sims"))

get_ds_dir(mod)

}
\seealso{
\code{\link[=mread_ds]{mread_ds()}}, \code{\link[=save_process_info]{save_process_info()}}, \code{\link[=move_ds]{move_ds()}}, \code{\link[=gc_ds]{gc_ds()}}
}
