% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{set_output_ds}
\alias{set_output_ds}
\alias{get_output_ds}
\title{Set or get default output directory}
\usage{
set_output_ds(x, output_dir)

get_output_ds(x)
}
\arguments{
\item{x}{a model object.}

\item{output_dir}{path to output directory.}
}
\description{
Set or get default output directory
}
\examples{
mod <- mrgsolve::house()

path <- file.path(tempfile(), "foo")

mod <- set_output_ds(mod, path)

get_output_ds(mod)

}
