% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{set_tempdir_base}
\alias{set_tempdir_base}
\title{Override the directory under which \code{mrgsim.ds} creates its temporary directory}
\usage{
set_tempdir_base(path)
}
\arguments{
\item{path}{The name of an existing directory.}
}
\value{
The absolute path (invisibly) to the subdirectory under \code{path} that
\code{mrgsim.ds} will use as its temporary directory.
}
\description{
By default, \code{mrgsim.ds} stores its files under a dedicated temporary
directory within \code{\link[=tempdir]{tempdir()}}. Use \code{set_tempdir_base()} to specify an
alternative location to \code{\link[=tempdir]{tempdir()}}.

This is useful in scenarios where the work will be spread across machines
(e.g., via Slurm) that do not shares access to the file system on which
\code{\link[=tempdir]{tempdir()}} resides.
}
\details{
Although the specified directory may reside anywhere, it should be treated as
a temporary directory. Use \code{\link[=save_ds]{save_ds()}} or \code{\link[=move_ds]{move_ds()}} to move the files to
more permanent locations.

\code{mrgsim.ds} may delete the backing parquet files in its temporary directory
(see \code{\link[=gc_ds]{gc_ds()}}), it does not delete \code{path} or the temporary directory that it
creates within it. The caller is expected to manage this temporary directory,
including its removal.

This function aborts if it is called after any \code{mrgsim.ds} functionality that
relies of the temporary directory.
}
\seealso{
\code{\link[=list_temp]{list_temp()}}
}
