% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_parquet_ds}
\alias{write_parquet_ds}
\title{Write simulations to a single parquet file}
\usage{
write_parquet_ds(x, sink, ...)
}
\arguments{
\item{x}{an mrgsimsds object.}

\item{sink}{passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}

\item{...}{passed to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}
}
\description{
Use this function to escape the \code{mrgsim.ds} universe, writing all simulated
data to a single \code{.parquet} file of your choosing.
}
